requirements.txt
# 核心套件
numpy>=1.26.0
opencv-python>=4.8.0
pillow>=10.0.0

# YOLO
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# OCR 車牌辨識
easyocr>=1.7.0

# 資料庫
psycopg2-binary==2.9.7

# 工具套件
pyyaml==6.0.1
python-dotenv==1.0.0

# 網頁伺服器
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0

# 選用 - 人臉辨識 (需要 Visual Studio Build Tools)
# cmake
# dlib
# face-recognition==1.3.0
